epm play zoom: fix install on i586 (ALT bug 42405)

+++ repack.d/zoom.sh
@@ -33,17 +33,19 @@ for i in libQt5 libav libfdkaac libmpg123 libquazip libturbojpeg libicu ; do
 done
 
 epm assure patchelf || exit
-for i in $BUILDROOT/opt/zoom/{libicui18n.so,libicui18n.so.56,libicuuc.so,libicuuc.so.56} ; do
-    a= patchelf --set-rpath '$ORIGIN/' $i
+for i in $BUILDROOT/opt/zoom/{libicui18n.so,libicui18n.so.*,libicuuc.so,libicuuc.so.*} ; do
+    a= patchelf --set-rpath '$ORIGIN/' $i || continue
 done
 
 for i in $BUILDROOT/opt/zoom/zoom ; do
     a= patchelf --set-rpath "$PRODUCTDIR/cef" $i
 done
 
-# qt5-3d libqt5-3dquickscene2d
-rm -v $BUILDROOT/opt/zoom/QtQuick/Scene2D/libqtquickscene2dplugin.so $BUILDROOT/opt/zoom/QtQuick/Scene3D/libqtquickscene3dplugin.so
-subst 's|.*/opt/zoom/QtQuick/Scene2D/libqtquickscene2dplugin.so.*||' $SPEC
-subst 's|.*/opt/zoom/QtQuick/Scene3D/libqtquickscene3dplugin.so.*||' $SPEC
+if [ -d $BUILDROOT/opt/zoom/QtQuick/Scene2D ] ; then
+    # qt5-3d libqt5-3dquickscene2d
+    rm -v $BUILDROOT/opt/zoom/QtQuick/Scene2D/libqtquickscene2dplugin.so $BUILDROOT/opt/zoom/QtQuick/Scene3D/libqtquickscene3dplugin.so
+    subst 's|.*/opt/zoom/QtQuick/Scene2D/libqtquickscene2dplugin.so.*||' $SPEC
+    subst 's|.*/opt/zoom/QtQuick/Scene3D/libqtquickscene3dplugin.so.*||' $SPEC
+fi
 
 [ -e $PRODUCTDIR/cef/chrome-sandbox ] && fix_chrome_sandbox $PRODUCTDIR/cef/chrome-sandbox
